vim: Make it possible to run clangd in a docker container!
Run `start-clangd.sh` in the project root containing
`compile_commands.json` inside of the docker environment. Make sure
to open up the port to the host. Start Nvim in the same project root but
on the host, open a C++ file and attach to the Clangd server with
`<Leader>cpp`. :^)