tmux: Set zsh as default-command
I found this to be needed to fix a weird behavior I was having. Any new
terminal (split or new window etc) took a second to initialize, but not
when ran `tmux split-window "/bin/zsh"`. Didn't seem to work to set
`default-shell` either, so I guess that the default shell is run with
some extra commands which makes zsh run extra stuff..?

Either way, this patch makes a new shell pop up instantaneously.